Add support for simulation via CLI without project export

+++ asreview/simulation/cli.py
@@ -123,25 +123,8 @@ def _cli_simulate(argv):
     # change the verbosity
     _set_log_verbosity(args.verbose)
 
-    # check for state file extension
-    if args.state_file is None:
-        raise ValueError("Specify project file name (with .asreview extension).")
-
-    # do this check now and again when zipping.
-    if Path(args.state_file).exists():
+    if args.state_file and Path(args.state_file).exists():
         raise ValueError("Project path already exists.")
-
-    # create a project file
-    fp_tmp_simulation = Path(args.state_file).with_suffix(".asreview.tmp")
-
-    project = Project.create(
-        fp_tmp_simulation,
-        project_id=Path(args.state_file).stem,
-        project_mode="simulate",
-        project_name=Path(args.state_file).stem,
-        project_description="Simulation created via ASReview via "
-        "command line interface",
-    )
 
     # Get a name for the dataset
     if re.match(r"^([a-zA-Z0-9_-]+)\:([a-zA-Z0-9_-]+)$", args.dataset):
@@ -151,10 +134,6 @@ def _cli_simulate(argv):
         filename = Path(args.dataset).name
 
     as_data = load_dataset(args.dataset)
-    as_data.to_file(Path(fp_tmp_simulation, "data", filename))
-
-    # Update the project.json.
-    project.update_config(dataset_path=filename)
 
     # create a new settings object from arguments
     settings = ReviewSettings(
@@ -210,7 +189,6 @@ def _cli_simulate(argv):
     fm = feature_model.fit_transform(
         as_data.texts, as_data.headings, as_data.bodies, as_data.keywords
     )
-    project.add_feature_matrix(fm, feature_model)
 
     print("The following records are prior knowledge:\n")
     for record_id, _ in as_data.df.iloc[prior_idx].iterrows():
@@ -238,10 +216,31 @@ def _cli_simulate(argv):
         )
     sim.review()
 
-    project.add_review(settings=settings, state=sim, status="finished")
+    if args.state_file is not None:
+        # write all results to the project file
+        fp_tmp_simulation = Path(args.state_file).with_suffix(".asreview.tmp")
 
-    project.export(args.state_file)
-    shutil.rmtree(fp_tmp_simulation)
+        project = Project.create(
+            fp_tmp_simulation,
+            project_id=Path(args.state_file).stem,
+            project_mode="simulate",
+            project_name=Path(args.state_file).stem,
+            project_description="Simulation created via ASReview via "
+            "command line interface",
+        )
+
+        as_data.to_file(Path(fp_tmp_simulation, "data", filename))
+        project.update_config(dataset_path=filename)
+
+        project.add_feature_matrix(fm, feature_model)
+        project.add_review(settings=settings, state=sim, status="finished")
+
+        # export the project file
+        project.export(args.state_file)
+        shutil.rmtree(fp_tmp_simulation)
+
+    else:
+        print("\nTo store the results, use the -s option. E.g. -s my_sim.asreview")
 
 
 DESCRIPTION_SIMULATE = """
@@ -261,7 +260,7 @@ def _simulate_parser(prog="simulate", description=DESCRIPTION_SIMULATE):
         formatter_class=argparse.RawTextHelpFormatter,
     )
 
-    # Active learning parameters
+    # active learning parameters
     # File path to the data.
     parser.add_argument(
         "dataset",
